refactor: move widgets.py into pan3d_components
